requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
pinecone>=5.0.0
ollama>=0.1.0
sentence-transformers>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
httpx>=0.24.0